+++ betterAutostart/View/ChildPanel/editProfilePanel.cs
@@ -7,6 +7,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
+using betterAutostart.Common;
 using betterAutostart.Common.ProfileClasses;
 
 namespace betterAutostart
