+++ betterAutostart/View/ChildPanel/executableAppPanel.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Runtime.InteropServices;
+using betterAutostart.Common;
 using betterAutostart.Common.ProfileClasses;
 
 namespace betterAutostart
