+++ betterAutostart/View/ChildPanel/profilePanel.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Timers;
 using System.Windows.Forms;
+using betterAutostart.Common;
 using betterAutostart.Common.ProfileClasses;
 using Timer = System.Timers;
