+++ betterAutostart/View/ChildPanel/settings.cs
@@ -8,6 +8,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Windows.Forms;
 using System.Diagnostics;
+using betterAutostart.Common;
 
 namespace betterAutostart
 {
