+++ betterAutostart/View/application.cs
@@ -2,6 +2,7 @@ using System;
 using System.Drawing;
 using System.Windows.Forms;
 using System.Runtime.InteropServices;
+using betterAutostart.Common;
 using NHotkey.WindowsForms;
 using NHotkey;
